Set the node image to node v18 instead of LTS
Although it is nice to use LTS, it is really annoying when the docker build
fails in CD due to an incompability introduced by lts targeting a different
node version.

Rather than having this fail in the commit, this requires us to manually
update the node version in the image. A little more tedium in exchange for
fewer surprises.